backend_examples/python/requirements.txt
supabase==2.3.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
typing-extensions==4.8.0